primeiro commit

+++ controllers/MonitoriaController.php
@@ -12,9 +12,12 @@ use app\models\Disciplina;
 use app\models\DisciplinaSearch;
 use yii\helpers\ArrayHelper;
 use yii\db\Command;
-use app\models\CursoSearch;
 use yii\filters\AccessControl;
 use yii\web\UploadedFile;
+use app\models\PeriodoInscricaoMonitoria;
+use app\models\DisciplinaPeriodo;
+use app\models\DisciplinaPeriodoSearch;
+use app\models\Aluno;
 
 /**
  * MonitoriaController implements the CRUD actions for Monitoria model.
@@ -26,10 +29,10 @@ class MonitoriaController extends Controller
         return [
             'acess' => [
                 'class' => AccessControl::className(),
-                'only' => ['create','index','update', 'view', 'delete'],
+                'only' => ['create','index','update', 'view', 'delete', 'minhasinscricoes', 'pendencias'],
                 'rules' => [
                     [
-                        'actions' => ['create','index','update', 'view', 'delete'],
+                        'actions' => ['create','index','update', 'view', 'delete', 'minhasinscricoes', 'pendencias'],
                         'allow' => true,
                         'matchCallback' => function ($rule, $action) 
                         {
@@ -59,6 +62,7 @@ class MonitoriaController extends Controller
      * Lists all Monitoria models.
      * @return mixed
      */
+    private $marcador = 1;
     public function actionIndex()
     {
         $searchModel = new MonitoriaSearch();
@@ -98,10 +102,16 @@ class MonitoriaController extends Controller
             $model->file->saveAs('uploads/historicos/'.$model->file->baseName.'.'.$model->file->extension);
             $model->pathArqHistorico = $model->file->name;
             $model->file = 'uploads/historicos/'.$model->file->baseName.'.'.$model->file->extension;
-			$model->IDAluno = Yii::$app->user->identity->id;
 
-            if ($model->save()) {
-				return $this->redirect(['view', 'id' => $model->ID]);
+            if ($model->validate()) {
+                //Número do Processo
+                $model->numProcs = date("Y").'/'.str_pad(strval($proxProcesso = Monitoria::find()->count() + 1), 6, '0', STR_PAD_LEFT);
+            }
+
+            if ($model->save()) 
+            {
+                return $this->redirect(['view', 'id' => $model->ID]);
+
             } else {
 
                 if ($model->errors) {
@@ -116,19 +126,29 @@ class MonitoriaController extends Controller
             }
         } else {
             //Número do Processo
-            $model->numProcs = date("Y").'/'.str_pad(strval($proxProcesso = Monitoria::find()->count() + 1), 6, '0', STR_PAD_LEFT);
-            //ID Aluno
-           // $model->IDAluno = 20902175;
+            $model->numProcs = '[Novo]';
+
+            //Aluno - Pega aluno baseando-se no CPF do usuário logado
+            $aluno = Aluno::findOne(['CPF' => Yii::$app->user->identity->login]);
+            $model->IDAluno = $aluno->ID;
+
+            //Status - Aguardando Avaliação
+            $model->status = 0;
+
+            //Seleciona o último período de inscrição
+            $periodoInscricao = PeriodoInscricaoMonitoria::find()->orderBy(['ID' => SORT_DESC])->one();
+            $model->IDperiodoinscr = $periodoInscricao->ID;
+            $periodo = $periodoInscricao->ano.'/'.$periodoInscricao->periodo;
 
             return $this->render('create', [
                 'model' => $model,
+                'periodo' => $periodo,
+                'matricula' => $aluno->matricula,
+                'banco' => $aluno->banco,
+                'agencia' => $aluno->agencia,
+                'conta' => $aluno->conta,
             ]);
         }
-    }
-
-    public function convert_multi_array($array) {
-      $out = implode("&",array_map(function($a) {return implode("~",$a);},$array));
-      return $out;
     }
 
     /**
@@ -139,15 +159,32 @@ class MonitoriaController extends Controller
      */
     public function actionUpdate($id)
     {
-        $model = $this->findModel($id);
+        if ( (Yii::$app->request->referrer) != '/monitoria/minhasinscricoes')
+        {
+            $model = $this->findModel($id);
 
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->ID]);
-        } else {
-            return $this->render('update', [
-                'model' => $model,
-            ]);
+            if ($model->load(Yii::$app->request->post()) && $model->save() ) {
+                return $this->redirect(['view', 'id' => $model->ID]);
+            } else {
+
+                //Aluno - Pega aluno baseando-se no CPF do usuário logado
+                $aluno = Aluno::findOne(['CPF' => Yii::$app->user->identity->login]);
+
+                //Seleciona o último período de inscrição
+                $periodoInscricao = PeriodoInscricaoMonitoria::find()->orderBy(['ID' => SORT_DESC])->one();
+                $periodo = $periodoInscricao->ano.'/'.$periodoInscricao->periodo;
+
+                return $this->render('update', [
+                    'model' => $model,
+                    'periodo' => $periodo,
+                    'matricula' => $aluno->matricula,
+                    'banco' => $aluno->banco,
+                    'agencia' => $aluno->agencia,
+                    'conta' => $aluno->conta,
+                ]);
+            }
         }
+        return $this->redirect(Yii::$app->request->referrer);
     }
 
     /**
@@ -158,21 +195,63 @@ class MonitoriaController extends Controller
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
+        if (Yii::$app->user->identity->perfil == 1) {
+            $this->findModel($id)->delete();
+            //$this->redirect(['index']);
+        } 
 
-        return $this->redirect(['index']);
+        return $this->redirect(Yii::$app->request->referrer);
     }
 
-    public function actionAcompanharmonitoria()
+    /**
+     * Finds the Monitoria model based on its primary key value.
+     * If the model is not found, a 404 HTTP exception will be thrown.
+     * @param integer $id
+     * @return Monitoria the loaded model
+     * @throws NotFoundHttpException if the model cannot be found
+     */
+    protected function findModel($id)
+    {
+        if (($model = Monitoria::findOne($id)) !== null) {
+            return $model;
+        } else {
+            throw new NotFoundHttpException('A página requisitada não existe.');
+        }
+    }
+
+    public function actionMinhasinscricoes()
     {
         $searchModel = new MonitoriaSearch();
-        $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
+        $dataProvider = $searchModel->searchAluno(Yii::$app->request->queryParams);
 
-        return $this->render('acompanharmonitoria', [
+        return $this->render('minhasinscricoes', [
             'searchModel' => $searchModel,
             'dataProvider' => $dataProvider,
         ]);
     }
+
+    public function actionPendencias() 
+    {
+        $searchModel = new MonitoriaSearch();
+        $dataProvider = $searchModel->searchPendencias();
+
+        return $this->render('pendencias', [
+            'searchModel' => $searchModel,
+            'dataProvider' => $dataProvider,
+        ]);
+    }
+    public function actionStatus($model)
+    {
+        if (Yii::$app->user->identity->perfil == 1) 
+        {
+            if ($model->load(Yii::$app->request->post()) && $model->save() ) {
+                return $this->redirect(['view', 'id' => $model->ID]);
+            }
+            else $this->render('status', ['model' => $model]);
+        } 
+
+        return $this->redirect(Yii::$app->request->referrer);
+    }   
 
     public function actionFazerplanosemestral()
     {
@@ -210,19 +289,8 @@ class MonitoriaController extends Controller
         return $this->render('gerarrelatorioanual');
     }
 
-    /**
-     * Finds the Monitoria model based on its primary key value.
-     * If the model is not found, a 404 HTTP exception will be thrown.
-     * @param integer $id
-     * @return Monitoria the loaded model
-     * @throws NotFoundHttpException if the model cannot be found
-     */
-    protected function findModel($id)
-    {
-        if (($model = Monitoria::findOne($id)) !== null) {
-            return $model;
-        } else {
-            throw new NotFoundHttpException('The requested page does not exist.');
-        }
+    public function convert_multi_array($array) {
+      $out = implode("&",array_map(function($a) {return implode("~",$a);},$array));
+      return $out;
     }
 }
